feat: Check collision against pipes
The pipe sprites are now part of the 'collisionObjects' array passed
down to the Player object, and the 'onCollision' method has been updated to
check if the player has collided against any of them.

If the player invades the space occupied by any of the pipes, whether at
the top or at the bottom of the screen, then the bird dies.